Rename pow.cpp/h -> poa.cpp/h
