+++ contrib/dapscoin-qt.pro
@@ -102,7 +102,7 @@ HEADERS += src/activemasternode.h \
            src/net.h \
            src/netbase.h \
            src/noui.h \
-           src/pow.h \
+           src/poa.h \
            src/protocol.h \
            src/pubkey.h \
            src/random.h \
@@ -406,7 +406,7 @@ SOURCES += src/activemasternode.cpp \
            src/net.cpp \
            src/netbase.cpp \
            src/noui.cpp \
-           src/pow.cpp \
+           src/poa.cpp \
            src/protocol.cpp \
            src/pubkey.cpp \
            src/random.cpp \
