+++ src/chain.h
@@ -8,7 +8,7 @@
 #ifndef BITCOIN_CHAIN_H
 #define BITCOIN_CHAIN_H
 
-#include "pow.h"
+#include "poa.h"
 #include "primitives/block.h"
 #include "tinyformat.h"
 #include "uint256.h"
