+++ src/main.cpp
@@ -22,7 +22,7 @@
 #include "net.h"
 #include "obfuscation.h"
 #include "poa.h"
-#include "pow.h"
+#include "poa.h"
 #include "swifttx.h"
 #include "txdb.h"
 #include "txmempool.h"
