+++ src/main.h
@@ -18,7 +18,7 @@
 #include "chainparams.h"
 #include "coins.h"
 #include "net.h"
-#include "pow.h"
+#include "poa.h"
 #include "primitives/block.h"
 #include "primitives/transaction.h"
 #include "script/script.h"
