+++ src/rpcmining.cpp
@@ -14,7 +14,7 @@
 #include "main.h"
 #include "miner.h"
 #include "net.h"
-#include "pow.h"
+#include "poa.h"
 #include "rpcserver.h"
 #include "util.h"
 #ifdef ENABLE_WALLET
