+++ src/txdb.cpp
@@ -8,7 +8,7 @@
 #include "txdb.h"
 
 #include "main.h"
-#include "pow.h"
+#include "poa.h"
 #include "uint256.h"
 
 #include <stdint.h>
